feat: ParticipantGrid 컴포넌트 분리
Task: #98
